More refactoring and fix tests
