+++ packages/cli/test/unit/cmds/beacon.test.ts
@@ -1,16 +1,16 @@
 import path from "node:path";
 import fs from "node:fs";
 import {expect} from "chai";
-import {createSecp256k1PeerId} from "@libp2p/peer-id-factory";
+import {createFromJSON, createSecp256k1PeerId} from "@libp2p/peer-id-factory";
 import {chainConfig} from "@lodestar/config/default";
 import {chainConfigToJson} from "@lodestar/config";
-import {createKeypairFromPeerId, ENR, SignableENR} from "@chainsafe/discv5";
+import {createKeypairFromPeerId, SignableENR} from "@chainsafe/discv5";
 import {exportToJSON} from "../../../src/config/peerId.js";
-import {beaconHandlerInit, initLogger, initPeerIdAndEnr} from "../../../src/cmds/beacon/handler.js";
+import {beaconHandlerInit} from "../../../src/cmds/beacon/handler.js";
+import {initPeerIdAndEnr} from "../../../src/cmds/beacon/initPeerIdAndEnr.js";
 import {BeaconArgs} from "../../../src/cmds/beacon/options.js";
 import {GlobalArgs} from "../../../src/options/globalOptions.js";
-import {testFilesDir} from "../../utils.js";
-import {getBeaconConfigFromArgs} from "../../../src/config/beaconParams.js";
+import {testFilesDir, testLogger} from "../../utils.js";
 
 describe("cmds / beacon / args handler", () => {
   // Make tests faster skipping a network call
@@ -35,14 +35,13 @@ describe("cmds / beacon / args handler", () => {
     const enrIp = "10.20.30.40";
     const enrTcp = 4000;
 
-    const {beaconPaths} = await runBeaconHandlerInit({
+    const {options} = await runBeaconHandlerInit({
       listenAddress: "0.0.0.0",
       "enr.ip": enrIp,
       "enr.tcp": enrTcp,
     });
 
-    const enrTxt = fs.readFileSync(path.join(beaconPaths.beaconDir, "enr"), "utf8");
-    const enr = ENR.decodeTxt(enrTxt);
+    const enr = options.network.discv5?.enr as SignableENR;
 
     expect(enr.ip).to.equal(enrIp, "wrong enr.ip");
     expect(enr.tcp).to.equal(enrTcp, "wrong enr.tcp");
@@ -58,14 +57,14 @@ describe("cmds / beacon / args handler", () => {
   it("Re-use existing peer", async () => {
     const prevPeerId = await createSecp256k1PeerId();
 
-    const peerIdFile = path.join(testFilesDir, "prev_peerid.json");
+    const peerIdFile = path.join(testFilesDir, "peer-id.json");
     fs.writeFileSync(peerIdFile, JSON.stringify(exportToJSON(prevPeerId)));
     const enr = SignableENR.createV4(createKeypairFromPeerId(prevPeerId));
     const enrFilePath = path.join(testFilesDir, "enr");
     fs.writeFileSync(enrFilePath, enr.encodeTxt());
 
     const {peerId} = await runBeaconHandlerInit({
-      peerIdFile,
+      persistNetworkIdentity: true,
     });
 
     expect(peerId.toString()).equal(prevPeerId.toString(), "peer must be equal to persisted");
@@ -98,72 +97,73 @@ describe("cmds / beacon / args handler", () => {
 });
 
 describe("initPeerIdAndEnr", () => {
-  it("should not reuse peer id, no peerIdFile", async () => {
-    const {peerId: peerId1} = await runBeaconHandlerInit({});
-    const {peerId: peerId2} = await runBeaconHandlerInit({});
+  it("should not reuse peer id, persistNetworkIdentity=false", async () => {
+    const {peerId: peerId1} = await initPeerIdAndEnr(
+      {persistNetworkIdentity: false} as BeaconArgs,
+      testFilesDir,
+      testLogger()
+    );
+    const {peerId: peerId2} = await initPeerIdAndEnr(
+      {persistNetworkIdentity: false} as BeaconArgs,
+      testFilesDir,
+      testLogger()
+    );
 
     expect(peerId1.toString()).not.equal(peerId2.toString(), "peer ids must be different");
   });
 
-  it("should not reuse peer id, not existed peerIdFile", async () => {
-    const notExistPeerIdFile = path.join(testFilesDir, "peer_id_not_exist.json");
-    const {peerId: peerId1} = await runBeaconHandlerInit({peerIdFile: notExistPeerIdFile});
-    const {peerId: peerId2} = await runBeaconHandlerInit({peerIdFile: notExistPeerIdFile});
+  it("should reuse peer id, persistNetworkIdentity=true", async () => {
+    const {peerId: peerId1} = await initPeerIdAndEnr(
+      {persistNetworkIdentity: true} as BeaconArgs,
+      testFilesDir,
+      testLogger()
+    );
+    const {peerId: peerId2} = await initPeerIdAndEnr(
+      {persistNetworkIdentity: true} as BeaconArgs,
+      testFilesDir,
+      testLogger()
+    );
 
-    expect(peerId1.toString()).not.equal(peerId2.toString(), "peer ids must be different");
+    expect(peerId1.toString()).to.equal(peerId2.toString(), "peer ids must be equal");
   });
 
-  it("should not reuse peer id, not able to load peerIdFile", async () => {
-    const peerIdFile = path.join(testFilesDir, "prev_peerid.json");
-    fs.writeFileSync(peerIdFile, "wrong peer id file content");
-    const {peerId: peerId1} = await runBeaconHandlerInit({peerIdFile});
-    const {peerId: peerId2} = await runBeaconHandlerInit({peerIdFile});
+  it("should overwrite invalid peer id", async () => {
+    const peerIdFile = path.join(testFilesDir, "peer-id.json");
+    const peerId1Str = "wrong peer id file content";
+    fs.writeFileSync(peerIdFile, peerId1Str);
+    const {peerId: peerId2} = await initPeerIdAndEnr(
+      {persistNetworkIdentity: true} as BeaconArgs,
+      testFilesDir,
+      testLogger()
+    );
+    const filePeerId = await createFromJSON(JSON.parse(fs.readFileSync(peerIdFile, "utf-8")));
 
-    expect(peerId1.toString()).not.equal(peerId2.toString(), "peer ids must be different");
+    expect(peerId1Str).not.equal(peerId2.toString(), "peer ids must be different");
+    expect(filePeerId.toString()).to.equal(peerId2.toString(), "peer ids must be equal");
   });
 
-  it("should not reuse peer id, not able to load enr file", async () => {
-    const prevPeerId = await createSecp256k1PeerId();
-    const peerIdFile = path.join(testFilesDir, "prev_peerid.json");
-    fs.writeFileSync(peerIdFile, JSON.stringify(exportToJSON(prevPeerId)));
-
+  it("should overwrite invalid enr", async () => {
     const enrFilePath = path.join(testFilesDir, "enr");
-    fs.writeFileSync(enrFilePath, "wrong enr file content");
+    const invalidEnr = "wrong enr file content";
+    fs.writeFileSync(enrFilePath, invalidEnr);
 
-    const {peerId: peerId1} = await runBeaconHandlerInit({peerIdFile});
-    const {peerId: peerId2} = await runBeaconHandlerInit({peerIdFile});
+    await initPeerIdAndEnr({persistNetworkIdentity: true} as BeaconArgs, testFilesDir, testLogger());
 
-    expect(peerId1.toString()).not.equal(peerId2.toString(), "peer ids must be different");
+    const validEnr = fs.readFileSync(enrFilePath, "utf-8");
+
+    expect(validEnr).not.equal(invalidEnr, "enrs must be different");
   });
 
-  it("should not reuse peer id, enr peer id is not the same to peer id file", async () => {
-    const prevPeerId = await createSecp256k1PeerId();
-    const peerIdFile = path.join(testFilesDir, "prev_peerid.json");
-    fs.writeFileSync(peerIdFile, JSON.stringify(exportToJSON(prevPeerId)));
-
-    const enrPeerId = await createSecp256k1PeerId();
-    const enr = SignableENR.createV4(createKeypairFromPeerId(enrPeerId));
+  it("should overwrite enr that doesn't match peer id", async () => {
+    const otherPeerId = await createSecp256k1PeerId();
+    const otherEnr = SignableENR.createFromPeerId(otherPeerId);
     const enrFilePath = path.join(testFilesDir, "enr");
-    fs.writeFileSync(enrFilePath, enr.encodeTxt());
+    const otherEnrStr = otherEnr.encodeTxt();
+    fs.writeFileSync(enrFilePath, otherEnrStr);
 
-    const {peerId} = await runInitPeerIdAndEnr({peerIdFile});
-    expect(peerId.toString()).not.equal(prevPeerId.toString(), "peer ids must be different");
-    expect(peerId.toString()).not.equal(enrPeerId.toString(), "peer ids must be different");
-  });
+    const {enr} = await initPeerIdAndEnr({persistNetworkIdentity: true} as BeaconArgs, testFilesDir, testLogger());
 
-  it("should reuse peer id and enr", async () => {
-    const prevPeerId = await createSecp256k1PeerId();
-
-    const peerIdFile = path.join(testFilesDir, "prev_peerid.json");
-    fs.writeFileSync(peerIdFile, JSON.stringify(exportToJSON(prevPeerId)));
-    const enr = SignableENR.createV4(createKeypairFromPeerId(prevPeerId));
-    const enrFilePath = path.join(testFilesDir, "enr");
-    fs.writeFileSync(enrFilePath, enr.encodeTxt());
-    // try 2 times
-    for (let i = 0; i < 2; i++) {
-      const {peerId} = await runInitPeerIdAndEnr({peerIdFile});
-      expect(peerId.toString()).equal(prevPeerId.toString(), "peer must be equal to persisted");
-    }
+    expect(enr.nodeId).not.equal(otherEnr, "enrs must be different");
   });
 });
 
@@ -173,12 +173,4 @@ async function runBeaconHandlerInit(args: Partial<BeaconArgs & GlobalArgs>) {
     dataDir: testFilesDir,
     ...args,
   } as BeaconArgs & GlobalArgs);
-}
-
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-async function runInitPeerIdAndEnr(args: Partial<BeaconArgs & GlobalArgs>) {
-  const args2 = {...args, dataDir: testFilesDir} as BeaconArgs & GlobalArgs;
-  const {config} = getBeaconConfigFromArgs(args2);
-  const logger = initLogger(args2, testFilesDir, config);
-  return initPeerIdAndEnr(args2, testFilesDir, logger);
 }
